Extract CONVERSION FACTORS from SVDBS

Extract a list of conversion factors from the SVDBS SURVAN_CONVERSION_FACTORS table

Usage

get_conversion_factors(channel)

Arguments

channel

an Object inherited from DBIConnection-class. This object is used to connect to communicate with the database engine. (see connect_to_database)

Value

A list is returned:

data

containing the result of the executed sql statement

sql

containing the sql call

colNames

a vector of the table's column names

Reference

Use the data dictionary for field name explanations

See also

Examples

if (FALSE) { # \dontrun{
# extracts complete CONVERSION FACTORS table
channel <- dbutils::connect_to_database(server="serverName",uid="userName")
get_conversion_factors(channel)
} # }