Extract species specific LENGTH-WEIGHT COEFFICIENTS from SVDBS

Pulls the length-weight coefficients from SVDBS LENGTH_WEIGHT_COEFFICIENTS table These coefficients are described in NOAA Tech Memo NMFS-NE-171.

get_length_weight(channel)

Arguments

channel

an Object inherited from DBIConnection-class. This object is used to communicate with the database engine. (see connect_to_database)

Value

Returns a data.table (nx12).

SVSPP

A standard code which represents a species caught in a trawl or dredge. Refer to the SVDBS.SVSPECIES_LIST

SVLWEXP

The exponent of the length-weight equation, b.

SVLWCOEFF

The natural log of the coefficient of the length-weight equation, ln a.

See also