Extract SEX information from SVDBS

Extract a list of sex codes and descriptions from the SVDBS.SEX_CODES table

get_sex(channel)

Arguments

channel

an Object inherited from DBIConnection-class. This object is used to connect to communicate with the database engine. (see connect_to_database)

Value

A list is returned:

data

tibble containing the result of the executed $sql statement

sql

containing the sql call

colNames

a vector of the table's column names

Reference

Use the data dictionary for field name explanations

See also

Examples

if (FALSE) { # \dontrun{
# extracts complete sex table
channel <- dbutils::connect_to_database(server="serverName",uid="userName")
get_sex(channel)
} # }