Assigns points to polygon

Assign survey data (points, lat and lon) to designated regions (polygons) from a shape file.

Usage

post_strat(surveyData, areaPolygon, areaDescription, na.keep = F)

Arguments

surveyData

Data table. NEFSC survey data generated by get_survdat_data.R

areaPolygon

sf object or character string. Default = "NEFSC strata". The default option uses the survey strata shapefile bundled with the package. To use any other shapefile for stratification, the shapefile must be read in as an sf object and the areaDescription argument must be specified.

areaDescription

Character String. Column name from areaPolygon that contains the strata designations.

na.keep

Boolean. Logical value to indicate whether original strata names should be retained.

Value

Returns a surveyData data.table with one additional column labeled with the value of areaDescription

areaDescription

The name of the region (found in areaPolygon) that a record in surveyData is assigned to

See also

Examples

if (FALSE) { # \dontrun{
# Called internally in 'strat_prep' which is then used to calculate stratified mean biomass
} # }