survdat: Pull and process Survey data

Tools to aid in extraction of survey data from the svdbs database

Details

Pulls bottom trawl survey data from svdbs

Applies corrections to the data based on changes in gear and ships

Partitions the data based on user supplied geospatial data

Estimates biomass of species

Helper functions to explain species, maturity, strata,vessel, sex codes

Pulls length-weight data and/or coefficients

References

To learn more about using survdat, start with the vignette: browseVignettes(package="survdat") or click the index below

See also

Author

Maintainer: Andy Beet andrew.beet@NOAA.gov (ORCID)

Authors:

Sean Lucey Sean.Lucey@NOAA.gov