Calculate swept area biomass

Calculates the biomass

swept_area_biomass(
  data,
  areaPolygon,
  areaDescription,
  filterByArea = "all",
  filterBySeason,
  species = "all",
 
    crs = "+proj=lcc +lat_1=20 +lat_2=60 +lat_0=40 +lon_0=-72 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0",
  merge.sex = T,
  poststrat = F,
  q = NULL,
  a = 0.0384,
  tidy = F
)

Arguments

data
Data frame. Survey data pull from svdbs with applied conversion factors

areaPolygon
sf object. Defining the areas by which to estimate biomass.

areaDescription
Character string. The name of the column in areaPolygon that describes the area (eg. "EPU","STRATA")

filterByArea
Character Vector. Vecor of area names. Names found in the areaPolygon in areaDescription Column. Biomass estimates will be aggregated over all areas specified in filterByArea (Default = "all")

filterBySeason
Character Vector. Vector of season names. Names found in the SEASON column of data

species
Character Vector. Vector of species SVSPP codes.

merge.sex
Boolean. Merge sexed species such as dogfish. (Default = T)

poststrat
Boolean. Indicating whether the original strata design was used or not. Changes the calculation for variance. (Default = F)

q
Table of survey catchability with a column of group names and a column of catchabilities. If not provided, assumes a q of 1 for each group (Minimum swept area estimates).

a
The average swept area of the trawl. Default value is the swept area of a standard NOAA Ship Albatross IV tow.

tidy
Boolean. Return output in long format (Default = F)

Value

Examples

if (FALSE) {
# pull data and apply conversion corrections
data <- survdatData

# read in "strata" shapefile
area <- sf::st_read(dsn = system.file("extdata","strata.shp",package="survdat"),quiet=T)

# estimate swept area biomass in the SPRING in STRATA = 1220, 1240, 1260:1290,1360:1400 for species = 300:310
swa <-swept_area_biomass(data=data, areaPolygon = area, areaDescription="STRATA", filterByArea=c(1220, 1240, 1260:1290,1360:1400),filterBySeason = "SPRING", species=c(300:310))

}