Calculate swept area biomass

Calculates the biomass. Details found here ...

calc_swept_area(
  surveyData,
  areaPolygon = "NEFSC strata",
  areaDescription = "STRATA",
  filterByArea = "all",
  filterBySeason,
  groupDescription = "SVSPP",
  filterByGroup = "all",
  mergesexFlag = T,
  tidy = F,
  q = NULL,
  a = 0.0384
)

Arguments

surveyData
Data table. NEFSC survey data generated by get_survdat_data.R

areaPolygon
sf object or character string. Default = "NEFSC strata". The default option uses the survey strata shapefile bundled with the package. To use any other shapefile for stratification, the shapefile must be read in as an sf object and the areaDescription argument must be specified.

areaDescription
Character String. Column name from areaPolygon that contains the strata designations.

filterByArea
Numeric vector. Set of areas to subset from the areaDescription of the areaPolygon.

filterBySeason
Character string. Which seasons of the surveyData should be included. Choices include "SPRING", "FALL", or "all".

groupDescription
Character string. Column of prepData which contains the groups (e.g. "SVSPP") on which the means are based.

filterByGroup
Character or numeric vector. Set of groups to subset from groupDescription. The default "all" will calculate means for all groups.

mergesexFlag
Boolean. Logical value to merge sexed species such as dogfish.

tidy
Boolean. Return output in long format (Default = F).

q
Data frame. Table of survey catchabilities with a column corresponding to groupDescription and a column of catchabilities. If NULL, assumes a q of 1 for each groupDescription (Minimum swept area estimates).

a
Numeric. The average swept area of the trawl. Default value is the swept area of a standard NOAA Ship Albatross IV tow.

Value

See also

Examples

if (FALSE) {
# Pull data and apply conversion corrections
data <- get_survdat_data(channel)
# Calculate swept area biomass for specific survey strata for the SPRING season
calc_swept_area(surveyData=data$survdat, filterByArea=c(1220, 1240, 1260:1290,1360:1400),filterBySeason = "SPRING")

# Calculate stratified mean for area defined by EPU regions, for all seasons ("SPRING", "FALL") and return in Tidy format
# Read in EPU shapefile (loaded as part of the package)
area <- sf::st_read(dsn = system.file("extdata","EPU.shp",package="survdat"),quiet=T)
calc_swept_area(surveyData=data$survdat, areaPolygon=area, areaDescription="EPU", filterByArea="all",filterBySeason = "all",tidy=T)

}